Add explicit file for integration tests
