+++ screens/CreateEventScreen.tsx
@@ -12,18 +12,22 @@ import { EventColors } from "@lib/events/EventColors"
 import { useDependencyValue } from "@lib/dependencies"
 import { EditEventInput, eventsDependencyKey } from "@lib/events"
 import { useMutation } from "react-query"
+import { dayjs } from "@lib/dayjs"
 
 export type CreateEventScreenProps = {
   locationInfo?: EventFormLocationInfo
 }
 
-const initialEventFormValues = {
-  title: "",
-  description: "",
-  dateRange: new FixedDateRange(new Date(), new Date()),
-  color: EventColors.Red,
-  shouldHideAfterStartDate: false,
-  radiusMeters: 0
+const createInitialFormValues = () => {
+  const now = new Date()
+  return {
+    title: "",
+    description: "",
+    dateRange: new FixedDateRange(now, dayjs(now).add(1, "hour").toDate()),
+    color: EventColors.Red,
+    shouldHideAfterStartDate: false,
+    radiusMeters: 0
+  }
 }
 
 /**
@@ -36,7 +40,7 @@ const CreateEventScreen = ({ locationInfo }: CreateEventScreenProps) => {
   })
   return (
     <EventForm
-      initialValues={{ ...initialEventFormValues, locationInfo }}
+      initialValues={{ ...createInitialFormValues(), locationInfo }}
       onSubmit={createEventMutation.mutateAsync}
     >
       <EventFormSubmitButton label="Create Event" />
